Add files via upload
This file defines the Mongoose schema for the student records in the project. It specifies the structure of each student document, including fields like fullname, age, address, grades, and TotalMarks. The schema also sets data types, validations, and required fields to ensure data consistency. By using this schema with Mongoose, the application can easily perform CRUD operations on the student collection in MongoDB.